requirements.txt
# Core
python-dotenv==1.0.0
pydantic>=2.0
pydantic-settings>=2.0

# Google Gemini
google-generativeai>=0.3.0

# Embeddings
sentence-transformers>=2.2.2

# Vector Database
chromadb>=0.4.24

# PDF Processing
pypdf>=3.17.0

# Text Processing
numpy>=1.24.0

# Frontend
streamlit>=1.28.0

# Visualization (optional)
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Development
pytest>=7.4.0